+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 ### Enhancements
 * Add support for [JVM target](https://github.com/realm/realm-kotlin/issues/62) (currently only macos).
+* Added support for marking a field as indexed with `@Index`
 
 ### Fixed
 * None.